+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Initial release
 - Fix ESP8266 Flash BUG
 - Fix download code BUG
+- Add gitee mirror